+++ src/tasking.cc
@@ -6,23 +6,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include <cstring>
-#include <ctime>
-#include <cstdio>
-#include <cassert>
-
-#include <algorithm>
-#include <array>
-#include <deque>
-#include <mutex>
-#include <random>
-
-#include "debug.h"
-
-#include "tasking.h"
-#include "threads.h"
-#include "numa_support.h"
-
 // Configuration options for the tasking support of LOMP
 #define USE_LINKED_LIST_LIFO_TASK_POOL 0
 #define USE_RESTR_LINKED_LIST_LIFO_TASK_POOL 0
@@ -34,16 +17,39 @@
 #define USE_SINGLE_TASK_POOL 0
 #define USE_MULTI_TASK_POOL 1
 
-#define USE_RANDOM_STEALING 0
+#define USE_RANDOM_STEALING 1
 #if (LOMP_TARGET_LINUX)
 // This task stealing algorithm requires Linux support thread affinity.
 #define USE_NUMA_AWARE_RANDOM_STEALING 1
 #else
 // Fallback implementation if we're not on a Linux system.
-#define USE_ROUND_ROBIN_STEALING 1
+#define USE_ROUND_ROBIN_STEALING 0
 #endif
 
 #define DEBUG_TASKING 0
+
+#include <cstring>
+#if DEBUG_TASKING
+#include <cstdio>
+#endif
+#include <cassert>
+
+#include <algorithm>
+#include <array>
+#include <deque>
+#include <mutex>
+#if USE_RANDOM_STEALING
+#include <random>
+#endif
+
+#include "debug.h"
+
+#include "memory.h"
+#include "tasking.h"
+#include "threads.h"
+#if USE_NUMA_AWARE_RANDOM_STEALING
+#include "numa_support.h"
+#endif
 
 namespace lomp::Tasking {
 
@@ -52,7 +58,7 @@ struct TaskPoolDeque {
   TaskPoolDeque() : taskCount(0), pool() {}
 
   bool put(TaskDescriptor * task) {
-    const std::lock_guard<Lock> guard(lock);
+    const auto guard = std::lock_guard{lock};
     if (taskCount >= maxSize) {
 #if DEBUG_TASKING
       printf("put: task pool full, task=%p, #tasks=%ld, max size deque: %ld\n",
@@ -73,7 +79,7 @@ struct TaskPoolDeque {
 
   TaskDescriptor * get() {
     TaskDescriptor * task = nullptr;
-    const std::lock_guard<Lock> guard(lock);
+    const auto guard = std::lock_guard{lock};
     if (taskCount > 0) {
       task = pool.back();
       pool.pop_back();
@@ -87,7 +93,7 @@ struct TaskPoolDeque {
 
   TaskDescriptor * steal() {
     TaskDescriptor * task = nullptr;
-    const std::lock_guard<Lock> guard(lock);
+    const auto guard = std::lock_guard{lock};
     if (taskCount > 0) {
       task = pool.front();
       pool.pop_front();
@@ -112,7 +118,7 @@ struct TaskPoolArrayLIFO {
   }
 
   bool put(TaskDescriptor * task) {
-    const std::lock_guard<Lock> guard(lock);
+    const auto guard = std::lock_guard{lock};
     if (taskCount >= maxSize) {
 #if DEBUG_TASKING
       printf("put: task pool full, task=%p, #tasks=%ld\n", task, taskCount);
@@ -130,7 +136,7 @@ struct TaskPoolArrayLIFO {
 
   TaskDescriptor * get() {
     TaskDescriptor * task = nullptr;
-    const std::lock_guard<Lock> guard(lock);
+    const auto guard = std::lock_guard{lock};
     if (taskCount > 0) {
       task = pool[taskCount - 1];
       pool[taskCount - 1] = nullptr;
@@ -161,12 +167,12 @@ struct TaskPoolLinkedListLIFO {
   TaskPoolLinkedListLIFO() : head(nullptr), taskCount(0) {}
 
   bool put(TaskDescriptor * task) {
-    auto node = new ListNode{nullptr, task};
+    auto * node = new ListNode{nullptr, task};
     if (!node) {
       return false;
     }
     {
-      const std::lock_guard<Lock> guard(lock);
+      const auto guard = std::lock_guard{lock};
       node->next = head;
       head = node;
       taskCount++;
@@ -182,14 +188,14 @@ struct TaskPoolLinkedListLIFO {
   TaskDescriptor * get() {
     ListNode * node = nullptr;
     TaskDescriptor * task = nullptr;
-    const std::lock_guard<Lock> guard(lock);
+    const auto guard = std::lock_guard{lock};
     if (head) {
       node = head;
       head = head->next;
     }
     if (node) {
       task = node->task;
-      delete node;
+      memory::delete_aligned_struct(node);
       taskCount--;
     }
     return task;
@@ -204,9 +210,14 @@ struct TaskPoolLinkedListLIFO {
 #endif
 
 private:
-  struct ListNode {
+  struct ListNode : private memory::CacheAligned {
+    ListNode(ListNode * next_, TaskDescriptor * task_) : next(next_), task(task_) {}
+
     ListNode * next;
     TaskDescriptor * task;
+
+    using CacheAligned::operator new;
+    using CacheAligned::operator delete;
   };
   ListNode * head;
   size_t taskCount;
@@ -221,7 +232,7 @@ struct TaskPoolRestrictedLinkedListLIFO {
   }
 
   bool put(TaskDescriptor * task) {
-    ListNode * node = new ListNode;
+    auto * node = new ListNode{nullptr, task};
     std::lock_guard<Lock> lock_guard(lock);
     node->task = task;
     node->next = nullptr;
@@ -257,7 +268,7 @@ struct TaskPoolRestrictedLinkedListLIFO {
     }
     if (node) {
       task = node->task;
-      delete node;
+      memory::delete_aligned_struct(node);
       taskCount--;
     }
     return task;
@@ -272,9 +283,14 @@ struct TaskPoolRestrictedLinkedListLIFO {
 #endif
 
 private:
-  struct ListNode {
+  struct ListNode : private memory::CacheAligned {
+    ListNode(ListNode * next_, TaskDescriptor * task_) : next(next_), task(task_) {}
+
     ListNode * next;
     TaskDescriptor * task;
+
+    using CacheAligned::operator new;
+    using CacheAligned::operator delete;
   };
   ListNode * head;
   size_t taskCount;
@@ -310,7 +326,7 @@ TaskPool * TaskPoolFactory() {
   pool = &taskPool;
 #endif
 #if USE_MULTI_TASK_POOL
-  pool = new TaskPool{};
+  pool = memory::make_aligned_struct<TaskPool>();
 #endif
 #if DEBUG_TASKING
   printf("task pool factory, task pool %p\n", pool);
@@ -343,9 +359,9 @@ size_t ComputeAllocSize(size_t sizeOfTaskClosure, size_t sizeOfShareds) {
 
 TaskDescriptor * AllocateTask(size_t sizeOfTaskClosure, size_t sizeOfShareds) {
   auto allocSize = ComputeAllocSize(sizeOfTaskClosure, sizeOfShareds);
-  auto task = static_cast<TaskDescriptor *>(malloc(allocSize));
+  auto * task = static_cast<TaskDescriptor *>(memory::make_aligned_chunk(allocSize));
   if (!task) {
-    // TODO: handle this error
+    lomp::fatalError("Could not allocate %d bytes for task descriptor", allocSize);
   }
 #if DEBUG_TASKING
   memset(static_cast<void *>(task), 0, allocSize);
@@ -358,8 +374,8 @@ TaskDescriptor * AllocateTask(size_t sizeOfTaskClosure, size_t sizeOfShareds) {
 
 void InitializeTaskDescriptor(TaskDescriptor * task, size_t sizeOfTaskClosure,
                               size_t sizeOfShareds, ThunkPointer task_entry) {
-  auto thread = Thread::getCurrentThread();
-  auto taskgroup = thread->getCurrentTaskgroup();
+  auto * thread = Thread::getCurrentThread();
+  auto * taskgroup = thread->getCurrentTaskgroup();
 
   // if task-private variables have been allocated, record the location
   // of the task-private space in the task descriptor
@@ -396,8 +412,8 @@ void InitializeTaskDescriptor(TaskDescriptor * task, size_t sizeOfTaskClosure,
 }
 
 void PrepareTask(TaskDescriptor * task) {
-  auto thread = Thread::getCurrentThread();
-  auto team = thread->getTeam();
+  auto * thread = Thread::getCurrentThread();
+  auto * team = thread->getTeam();
 
   // Count this task as being created for determining how many tasks are left to
   // be executed.
@@ -418,14 +434,14 @@ void PrepareTask(TaskDescriptor * task) {
 
   // Now we have to also record this task as active for a potentially active
   // taskgroup
-  if (auto taskgroup = task->metadata.taskgroup; taskgroup) {
+  if (auto * taskgroup = task->metadata.taskgroup; taskgroup) {
     assert(taskgroup->activeTasks.load() >= 0);
     taskgroup->activeTasks++;
   }
 }
 
 bool StoreTask(TaskDescriptor * task) {
-  auto taskPool = Thread::getCurrentThread()->getTaskPool();
+  auto * taskPool = Thread::getCurrentThread()->getTaskPool();
 
   // Try to put the task into the pool.
   if (taskPool->put(task)) {
@@ -441,18 +457,18 @@ bool StoreTask(TaskDescriptor * task) {
 
 void FreeTask(TaskDescriptor * task) {
   // memset(task, 0, sizeof(TaskDescriptor));
-  free(task);
+  memory::delete_aligned_chunk(task);
 }
 
 void FreeTaskAndAncestors(TaskDescriptor * task) {
   // This lock prevents freeing tasks while another thread is also attempting to
   // purge completed tasks.
   static std::mutex lock;
-  std::lock_guard<std::mutex> guard(lock);
+  const auto guard = std::lock_guard{lock};
   size_t children = 0;
 
 #if DEBUG_TASKING
-  auto thread = Thread::getCurrentThread();
+  auto * thread = Thread::getCurrentThread();
   int tid = thread->getLocalId();
 #endif
 
@@ -483,8 +499,8 @@ void FreeTaskAndAncestors(TaskDescriptor * task) {
 }
 
 void InvokeTask(TaskDescriptor * task) {
-  auto thread = Thread::getCurrentThread();
-  auto team = thread->getTeam();
+  auto * thread = Thread::getCurrentThread();
+  auto * team = thread->getTeam();
   int32_t gtid = 0;
 
   // store the reference to the previously running task
@@ -530,8 +546,8 @@ void InvokeTask(TaskDescriptor * task) {
 }
 
 void CompleteTask(TaskDescriptor * task) {
-  auto thread = Thread::getCurrentThread();
-  auto team = thread->getTeam();
+  auto * thread = Thread::getCurrentThread();
+  auto * team = thread->getTeam();
 
   task->metadata.flags = TaskDescriptor::Flags::Completed;
 
@@ -551,7 +567,7 @@ void CompleteTask(TaskDescriptor * task) {
 
   // Now we have to also record this task as being no longer active for a
   // potentially active taskgroup
-  if (auto taskgroup = task->metadata.taskgroup; taskgroup) {
+  if (auto * taskgroup = task->metadata.taskgroup; taskgroup) {
     --taskgroup->activeTasks;
     assert(taskgroup->activeTasks.load() >= 0);
   }
@@ -565,8 +581,8 @@ void CompleteTask(TaskDescriptor * task) {
 #if USE_ROUND_ROBIN_STEALING
 struct RoundRobinStealTask {
   TaskDescriptor * operator()() {
-    auto thread = Thread::getCurrentThread();
-    auto team = thread->getTeam();
+    auto * thread = Thread::getCurrentThread();
+    auto * team = thread->getTeam();
     auto me = thread->getLocalId();
     auto teamSize = team->getCount();
     TaskDescriptor * task = nullptr;
@@ -590,8 +606,8 @@ struct RoundRobinStealTask {
 #if USE_RANDOM_STEALING
 struct RandomStealTask {
   TaskDescriptor * operator()() {
-    auto thread = Thread::getCurrentThread();
-    auto team = thread->getTeam();
+    auto * thread = Thread::getCurrentThread();
+    auto * team = thread->getTeam();
     auto me = thread->getLocalId();
     auto teamSize = team->getCount();
     TaskDescriptor * task = nullptr;
@@ -609,7 +625,7 @@ struct RandomStealTask {
 #if USE_NUMA_AWARE_RANDOM_STEALING
 struct NumaStealStask {
   TaskDescriptor * operator()() {
-    auto thread = Thread::getCurrentThread();
+    auto * thread = Thread::getCurrentThread();
     TaskDescriptor * task = nullptr;
 
     // Determine NUMA domain of the thief
@@ -677,8 +693,8 @@ struct StealTask : private StealTaskImpl {
 };
 
 bool ScheduleTask() {
-  auto thread = Thread::getCurrentThread();
-  auto taskPool = thread->getTaskPool();
+  auto * thread = Thread::getCurrentThread();
+  auto * taskPool = thread->getTaskPool();
   bool result = false;
 
   // Try to retrieve a task from the task pool.
@@ -704,8 +720,8 @@ bool ScheduleTask() {
 }
 
 void TaskExecutionBarrier(bool internalBarrier) {
-  auto thread = Thread::getCurrentThread();
-  auto team = thread->getTeam();
+  auto * thread = Thread::getCurrentThread();
+  auto * team = thread->getTeam();
   size_t teamSize = team->getCount();
   size_t goal = internalBarrier ? teamSize : 0;
 #if DEBUG_TASKING
@@ -724,7 +740,7 @@ void TaskExecutionBarrier(bool internalBarrier) {
 bool TaskWait() {
   // Determine the task that executes the encountered taskwait: it is the
   // currently scheduled task on the current threads.
-  auto thread = Thread::getCurrentThread();
+  auto * thread = Thread::getCurrentThread();
   TaskDescriptor * parent = thread->getCurrentTask();
 
   if (parent) {
@@ -767,10 +783,10 @@ bool TaskWait() {
 }
 
 void TaskgroupBegin() {
-  auto thread = Thread::getCurrentThread();
-  auto outer = thread->getCurrentTaskgroup();
+  auto * thread = Thread::getCurrentThread();
+  auto * outer = thread->getCurrentTaskgroup();
 
-  auto inner = new Taskgroup(outer);
+  auto * inner = memory::make_aligned_struct<Taskgroup>(outer);
   thread->setCurrentTaskgroup(inner);
 
 #if DEBUG_TASKING
@@ -780,8 +796,8 @@ void TaskgroupBegin() {
 }
 
 void TaskgroupEnd() {
-  auto thread = Thread::getCurrentThread();
-  auto taskgroup = thread->getCurrentTaskgroup();
+  auto * thread = Thread::getCurrentThread();
+  auto * taskgroup = thread->getCurrentTaskgroup();
 
 #if DEBUG_TASKING
   auto id = thread->getLocalId();
@@ -802,7 +818,7 @@ void TaskgroupEnd() {
       //   ;
     }
 
-    auto outer = taskgroup->outer;
+    auto * outer = taskgroup->outer;
     thread->setCurrentTaskgroup(outer);
   }
 }